ET_Speed_Data_Pipeline/load-nyctraffic-speed/requirements.txt
functions_framework==3.6.0
google-cloud-storage==2.16.0
python-dotenv==1.0.1
google-cloud-bigquery

